@@ -1,8 +1,17 @@
 {
     "name": "emilybache/gilded-rose-refactoring-kata",
     "description": "A kata to practice refactoring, tests and polymorphism",
+    "license": "MIT",
     "require": {
-        "php": "^7.3 || ^8.0"
+        "php": "^8.0"
+    },
+    "require-dev": {
+        "approvals/approval-tests": "dev-Main",
+        "phpunit/phpunit": "^9.5",
+        "phpstan/phpstan": "^1.9",
+        "phpstan/phpstan-phpunit": "^1.3",
+        "symplify/easy-coding-standard": "^11.1",
+        "symplify/phpstan-extensions": "^11.1"
     },
     "autoload": {
         "psr-4": {
@@ -14,18 +23,7 @@
             "Tests\\": "tests/"
         }
     },
-    "require-dev": {
-        "phpunit/phpunit": "^9.5",
-        "phpstan/phpstan": "^0.12.85",
-        "phpstan/phpstan-phpunit": "^0.12.18",
-        "symplify/easy-coding-standard": "^9.3",
-        "symplify/phpstan-extensions": "^9.3",
-        "approvals/approval-tests": "^1.4"
-    },
     "scripts": {
-        "checkcode": "phpcs src tests --standard=PSR12",
-        "fixcode": "phpcbf src tests --standard=PSR12",
-        "test": "phpunit",
         "tests": "phpunit",
         "test-coverage": "phpunit --coverage-html build/coverage",
         "check-cs": "ecs check",
