@@ -7,13 +7,11 @@ namespace GildedRose;
 final class GildedRose
 {
     /**
-     * @var Item[]
+     * @param Item[] $items
      */
-    private $items;
-
-    public function __construct(array $items)
-    {
-        $this->items = $items;
+    public function __construct(
+        private array $items
+    ) {
     }
 
     public function updateQuality(): void
@@ -29,12 +27,12 @@ final class GildedRose
                 if ($item->quality < 50) {
                     $item->quality = $item->quality + 1;
                     if ($item->name == 'Backstage passes to a TAFKAL80ETC concert') {
-                        if ($item->sell_in < 11) {
+                        if ($item->sellIn < 11) {
                             if ($item->quality < 50) {
                                 $item->quality = $item->quality + 1;
                             }
                         }
-                        if ($item->sell_in < 6) {
+                        if ($item->sellIn < 6) {
                             if ($item->quality < 50) {
                                 $item->quality = $item->quality + 1;
                             }
@@ -44,10 +42,10 @@ final class GildedRose
             }
 
             if ($item->name != 'Sulfuras, Hand of Ragnaros') {
-                $item->sell_in = $item->sell_in - 1;
+                $item->sellIn = $item->sellIn - 1;
             }
 
-            if ($item->sell_in < 0) {
+            if ($item->sellIn < 0) {
                 if ($item->name != 'Aged Brie') {
                     if ($item->name != 'Backstage passes to a TAFKAL80ETC concert') {
                         if ($item->quality > 0) {
