@@ -1,6 +1,6 @@
 package com.gildedrose
 
-class GildedRose(var items: Array<Item>) {
+class GildedRose(var items: List<Item>) {
 
     fun updateQuality() {
         for (i in items.indices) {
