@@ -4,7 +4,7 @@ fun main(args: Array<String>) {
 
     println("OMGHAI!")
 
-    val items = arrayOf(Item("+5 Dexterity Vest", 10, 20), //
+    val items = listOf(Item("+5 Dexterity Vest", 10, 20), //
             Item("Aged Brie", 2, 0), //
             Item("Elixir of the Mongoose", 5, 7), //
             Item("Sulfuras, Hand of Ragnaros", 0, 80), //
@@ -31,6 +31,4 @@ fun main(args: Array<String>) {
         println()
         app.updateQuality()
     }
-
-
 }
